requirements.txt

cryptography==41.0.1
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.5
